refactor(call-recorder): share Recall webhook metadata reader, tighten target typing/test
