+++ packages/twenty-apps/public/call-recorder/src/logic-functions/process-recall-webhook.ts
@@ -6,9 +6,7 @@ import { handleRecallWebhook } from 'src/logic-functions/flows/handle-recall-web
 import { type RecallWebhookBody } from 'src/logic-functions/recall-api/parse-recall-webhook-event.util';
 
 // Dispatched by the recall-webhook resolver; runs in the resolved workspace so the client is workspace-scoped.
-export const processRecallWebhookHandler = (
-  body: RecallWebhookBody,
-): Promise<object> =>
+export const processRecallWebhookHandler = (body: RecallWebhookBody) =>
   handleRecallWebhook({
     client: new CoreApiClient(),
     body,
