+++ packages/twenty-apps/public/call-recorder/src/logic-functions/recall-api/extract-twenty-workspace-id-from-recall-webhook.util.ts
@@ -1,18 +1,8 @@
+import { getRecallWebhookBotMetadata } from 'src/logic-functions/recall-api/get-recall-webhook-bot-metadata.util';
 import { type RecallWebhookBody } from 'src/logic-functions/recall-api/parse-recall-webhook-event.util';
-import { asRecord } from 'src/logic-functions/utils/as-record.util';
-import { getRecordAtPath } from 'src/logic-functions/utils/get-record-at-path.util';
 import { getString } from 'src/logic-functions/utils/get-string.util';
 
 export const extractTwentyWorkspaceIdFromRecallWebhook = (
   body: RecallWebhookBody,
-): string | undefined => {
-  const data = asRecord(body.data);
-  const bot = asRecord(body.bot);
-  const metadata =
-    asRecord(bot?.metadata) ??
-    asRecord(getRecordAtPath(data, ['bot', 'metadata'])) ??
-    asRecord(getRecordAtPath(data, ['recording', 'metadata'])) ??
-    asRecord(data?.metadata);
-
-  return getString(metadata?.twentyWorkspaceId);
-};
+): string | undefined =>
+  getString(getRecallWebhookBotMetadata(body)?.twentyWorkspaceId);
